refactor: move CategoryFIlterList to SharedComponents
