+++ src/components/SharedComponents/CategoryFilterList/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./CategoryFilterList";
+export type { LearningThemeSelectedTrackingProps } from "./CategoryFilterList";
