+++ src/components/TeacherComponents/UnitsLearningThemeFilters/index.ts
@@ -1,2 +1,1 @@
 export { default } from "./UnitsLearningThemeFilters";
-export type { LearningThemeSelectedTrackingProps } from "./UnitsLearningThemeFilters";
